moved node types to parser/nodes
