+++ tempe/parser/nodes/ifclause.go
@@ -1,4 +1,4 @@
-package parser
+package nodes
 
 import (
 	"fmt"
