+++ tempe/parser/nodes/ifstatement.go
@@ -1,4 +1,4 @@
-package parser
+package nodes
 
 import (
 	"fmt"
