+++ tempe/parser/nodes/statement.go
@@ -1,4 +1,4 @@
-package parser
+package nodes
 
 import (
 	"io"
