+++ tempe/parser/nodes/template_nodes_collection.go
@@ -1,4 +1,4 @@
-package parser
+package nodes
 
 import (
 	"io"
