+++ tempe/parser/nodes/text.go
@@ -1,4 +1,4 @@
-package parser
+package nodes
 
 import (
 	"fmt"
