+++ tempe/parser/nodes/whitespace.go
@@ -1,4 +1,4 @@
-package parser
+package nodes
 
 import (
 	"regexp"
